+++ peers/find_tag_match.js
@@ -1,11 +1,20 @@
+const {decodeChanId} = require('bolt07');
+
 const {shuffle} = require('./../arrays');
+const {isMatchingFilters} = require('./../display');
+
+const sumOf = arr => arr.reduce((sum, n) => sum + n, Number());
 
 /** Find a node for a tag query
 
   {
     channels: [{
+      id: <Standard Format Channel Id String>
+      local_balance: <Channel Local Balance Tokens Number>
       partner_public_key: <Peer Public Key Hex String>
+      remote_balance: <Channel Local Balance Tokens Number>
     }]
+    [filters]: [<Filter Expression String>]
     query: <Query String>
     tags: [{
       [alias]: <Tag Alias String>
@@ -16,6 +25,10 @@ const {shuffle} = require('./../arrays');
 
   @returns
   {
+    [failure]: {
+      error: <Error String>
+      formula: <Errored Formula String>
+    }
     [match]: <Matching Node Public Key Hex String>
     [matches]: [{
       [alias]: <Tag Alias String>
@@ -24,7 +37,7 @@ const {shuffle} = require('./../arrays');
     }]
   }
 */
-module.exports = ({channels, tags, query}) => {
+module.exports = ({channels, filters, tags, query}) => {
   const peerKeys = channels.map(n => n.partner_public_key);
 
   // Find tags that match on id or on alias, and also have relevant nodes
@@ -56,12 +69,52 @@ module.exports = ({channels, tags, query}) => {
   }
 
   // Get the array of nodes in the tag match
-  const array = tagMatch.nodes.filter(n => peerKeys.includes(n));
+  const tagMatches = tagMatch.nodes.filter(n => peerKeys.includes(n));
+
+  // Filter out matches in the array of peers that do not fulfill criteria
+  const array = tagMatches
+    .map(key => {
+      if (!filters || !filters.length) {
+        return {match: key};
+      }
+
+      const withPeer = channels.filter(n => n.partner_public_key === key);
+
+      const matching = isMatchingFilters({
+        filters: filters || [],
+        variables: {
+          heights: withPeer.map(n => decodeChanId({channel: n.id}).block_height),
+          inbound_liquidity: sumOf(withPeer.map(n => n.remote_balance)),
+          outbound_liquidity: sumOf(withPeer.map(n => n.local_balance)),
+        },
+      });
+
+      if (!!matching.failure) {
+        return matching;
+      }
+
+      if (!matching.is_matching) {
+        return;
+      }
+
+      return {match: key};
+    })
+    .filter(n => !!n);
+
+  // Exit early when there is no match
+  if (!array.length) {
+    return {};
+  }
+
+  // Exit early when there is a failure in the tag
+  if (!!array.find(n => !!n.failure)) {
+    return array.find(n => !!n.failure);
+  }
 
   // Shuffle the results
   const {shuffled} = shuffle({array});
 
-  const [match] = shuffled;
+  const [{match}] = shuffled;
 
   return {match};
 };