+++ test/peers/test_find_tag_match.js
@@ -4,7 +4,12 @@ const {findTagMatch} = require('./../../peers');
 
 const makeArgs = overrides => {
   const args = {
-    channels: [{partner_public_key: Buffer.alloc(33, 3).toString('hex')}],
+    channels: [{
+      id: '1x1x1',
+      local_balance: 1,
+      partner_public_key: Buffer.alloc(33, 3).toString('hex'),
+      remote_balance: 2,
+    }],
     query: '0000',
     tags: [{
       id: Buffer.alloc(32).toString('hex'),
@@ -24,6 +29,43 @@ const tests = [
     expected: {match: Buffer.alloc(33, 3).toString('hex')},
   },
   {
+    args: makeArgs({filters: ['invalid formula']}),
+    description: 'A failed formula is provided',
+    expected: {
+      failure: {error: 'FailedToParseFormula', formula: 'invalid formula'},
+    },
+  },
+  {
+    args: makeArgs({
+      channels: [
+        {
+          id: '1x1x1',
+          local_balance: 1e3,
+          partner_public_key: Buffer.alloc(33, 3).toString('hex'),
+          remote_balance: 1,
+        },
+        {
+          id: '2x2x2',
+          local_balance: 1e8,
+          partner_public_key: Buffer.alloc(33, 2).toString('hex'),
+          remote_balance: 1,
+        },
+      ],
+      filters: ['outbound_liquidity > 5*m'],
+      tags: [{
+        id: Buffer.alloc(32).toString('hex'),
+        nodes: [
+          Buffer.alloc(33, 2).toString('hex'),
+          Buffer.alloc(33, 3).toString('hex'),
+        ],
+      }],
+    }),
+    description: 'A tagged node is found which complies with a filter',
+    expected: {
+      match: Buffer.alloc(33, 2).toString('hex'),
+    },
+  },
+  {
     args: makeArgs({
       tags: [{id: Buffer.alloc(32).toString('hex')}],
     }),
@@ -33,8 +75,18 @@ const tests = [
   {
     args: makeArgs({
       channels: [
-        {partner_public_key: Buffer.alloc(33, 2).toString('hex')},
-        {partner_public_key: Buffer.alloc(33, 3).toString('hex')},
+        {
+          id: '1x1x1',
+          local_balance: 1,
+          partner_public_key: Buffer.alloc(33, 2).toString('hex'),
+          remote_balance: 1,
+        },
+        {
+          id: '1x1x1',
+          local_balaance: 1,
+          partner_public_key: Buffer.alloc(33, 3).toString('hex'),
+          remote_balance: 1,
+        },
       ],
       query: 'alias',
       tags: [